Fix lint plugin with kotlin 2.3.20
